requirements.txt
flask
opencv-python
numpy
mysql-connector-python
gunicorn
